// src/static/enums/env-names.enum.ts
export enum EnvNamesEnum {
  port = 'PORT',
  db_host = 'DATABASE_HOST',
  db_port = 'DATABASE_PORT',
  db_name = 'DATABASE_NAME',
  db_user = 'DATABASE_USER',
  db_password = 'DATABASE_PASSWORD',
  redis_enabled = 'REDIS_ENABLED',
  redis_host = 'REDIS_HOST',
  redis_port = 'REDIS_PORT',
  redis_password = 'REDIS_PASSWORD',
  secret_word = 'SECRET_WORD',
  expires_in = 'EXPIRES_IN',
}